+++ src/main/java/racingcar/domain/NumberSelector.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 public interface NumberSelector {
     int selectNumber();
