+++ src/test/java/racingcar/domain/MoveNumberSelector.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 public class MoveNumberSelector implements NumberSelector {
     @Override
