+++ src/test/java/racingcar/domain/RacingGameTest.java
@@ -1,12 +1,12 @@
-package racingcar;
+package racingcar.domain;
 
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
 import java.util.Optional;
 
+import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
-import static org.assertj.core.api.AssertionsForClassTypes.*;
 
 public class RacingGameTest {
